\name{pd.110405.cdiffusa.lz.exp}
\Rdversion{1.1}
\docType{package}
\alias{pd.110405.cdiffusa.lz.exp}
\title{
  Annotation package for pd.110405.cdiffusa.lz.exp.
}
\description{
  Annotation package for pd.110405.cdiffusa.lz.exp built with pdInfoBuilder.
}
\details{
  This package is to be used in conjunction with the oligo package.
}
\keyword{package}
