\name{data:pmSequence}
\Rdversion{1.1}
\alias{pmSequence}
\alias{data:pmSequence}
\alias{bgSequence}
\alias{data:bgSequence}
\alias{mmSequence}
\alias{data:mmSequence}
\docType{data}
\title{
  Sequence data for PM probes.
}
\description{
  Sequence data for PM probes.
}
\usage{
data(pmSequence)
}
\format{
Sequences are stored within an DataFrame object with two columns: 'fid'
and 'sequence'.
}
\examples{
data(pmSequence)
}
\keyword{datasets}
